Enhance code structure
